Refactored limited subset of tests into prioritized groupings to use for live pair programming assessment.
