+++ app/01-basics.js
@@ -1,0 +1,49 @@
+/* eslint-disable */
+exports = typeof window === 'undefined' ? global : window;
+
+/**
+ * This file defines an object with some methods. Some of these methods are
+ * populated incorrectly; your job is to fix them. Other methods are not
+ * populated at all; your job is to fill them out.
+ */
+
+// Best Practices
+exports.bestPracticesAnswers = {
+    globals: function() {
+        myObject = {
+            name: 'Jory'
+        };
+
+        return myObject;
+    },
+
+    comparison: function(val1, val2) {
+
+    }
+};
+
+// Logical Operators
+exports.logicalOperatorsAnswers = {
+    or: function(a, b) {
+
+    },
+
+    and: function(a, b) {
+
+    }
+};
+
+// Flow Control
+exports.flowControlAnswers = {
+    fizzBuzz: function(num) {
+        // write a function that receives a number as its argument;
+        // if the number is divisible by 3, the function should return 'fizz';
+        // if the number is divisible by 5, the function should return 'buzz';
+        // if the number is divisible by 3 and 5, the function should return
+        // 'fizzbuzz';
+        //
+        // otherwise the function should return the number, or false if no number
+        // was provided or the value provided is not a number
+
+    }
+};
