+++ app/02-arrays-objects.js
@@ -1,0 +1,33 @@
+exports = typeof window === 'undefined' ? global : window;
+
+exports.arraysAnswers = {
+  sum: function(arr) {
+
+  },
+
+  count: function(arr, item) {
+
+  },
+
+  remove: function(arr, item) {
+
+  },
+
+  duplicates: function(arr) {
+
+  },
+
+  findAllOccurrences: function(arr, target) {
+
+  }
+};
+
+exports.objectsAnswers = {
+  alterContext: function(fn, obj) {
+
+  },
+
+  iterate: function(obj) {
+
+  }
+};
