+++ app/04-concepts.js
@@ -1,0 +1,91 @@
+exports = typeof window === 'undefined' ? global : window;
+
+// Async
+exports.asyncAnswers = {
+  /**
+   * This test describes a function, count, that takes two arguments: a starting number,
+   * and an ending number. The function should console.log each number from the start
+   * number to the end number, one number per 1/10th of a second. The function should
+   * return an object with a cancel method, which should cancel the counting.
+   */
+  count: function (start, end) {
+
+  },
+
+  async: function(value) {
+
+  },
+
+  manipulateRemoteData: function(url) {
+
+  }
+};
+
+// Modules
+exports.modulesAnswers = {
+  createModule: function(str1, str2) {
+
+  }
+};
+
+// More Arrays
+exports.moreArraysAnswers = {
+  removeWithoutCopy: function(arr, item) {
+
+  }
+};
+
+// More Objects
+exports.moreObjectsAnswers = {
+  alterObjects: function (constructor, greeting) {
+
+  }
+};
+
+// More Functions
+exports.moreFunctionsAnswers = {
+  callIt: function(fn) {
+
+  },
+
+  partialUsingArguments: function(fn) {
+
+  },
+
+  curryIt: function(fn) {
+
+  }
+};
+
+// Recursion
+exports.recursionAnswers = {
+  listFiles: function(data, dirName) {
+
+  }
+};
+
+// Permutation
+exports.permutationAnswers = {
+  permute: function(arr) {
+
+  },
+
+  fibonacci: function(n) {
+
+  },
+
+  validParentheses: function(n) {
+
+  }
+};
+
+// Fun w/ Strings
+exports.stringsAnswers = {
+  reduceString: function(str, amount) {
+
+  },
+
+  wordWrap: function(str, cols) {
+
+  }
+};
