--- app/async.js
@@ -1,11 +1,0 @@
-exports = typeof window === 'undefined' ? global : window;
-
-exports.asyncAnswers = {
-  async: function(value) {
-
-  },
-
-  manipulateRemoteData: function(url) {
-
-  }
-};
