--- app/count.js
@@ -1,7 +1,0 @@
-exports = typeof window === 'undefined' ? global : window;
-
-exports.countAnswers = {
-  count: function (start, end) {
-
-  }
-};
