--- app/logicalOperators.js
@@ -1,11 +1,0 @@
-exports = typeof window === 'undefined' ? global : window;
-
-exports.logicalOperatorsAnswers = {
-  or: function(a, b) {
-
-  },
-
-  and: function(a, b) {
-
-  }
-};
