--- app/modules.js
@@ -1,7 +1,0 @@
-exports = typeof window === 'undefined' ? global : window;
-
-exports.modulesAnswers = {
-  createModule: function(str1, str2) {
-
-  }
-};
