--- app/numbers.js
@@ -1,19 +1,0 @@
-exports = typeof window === 'undefined' ? global : window;
-
-exports.numbersAnswers = {
-  valueAtBit: function(num, bit) {
-
-  },
-
-  base10: function(str) {
-
-  },
-
-  convertToBinary: function(num) {
-
-  },
-
-  multiply: function(a, b) {
-
-  }
-};
