--- app/objects.js
@@ -1,15 +1,0 @@
-exports = typeof window === 'undefined' ? global : window;
-
-exports.objectsAnswers = {
-  alterContext: function(fn, obj) {
-
-  },
-
-  alterObjects: function(constructor, greeting) {
-
-  },
-
-  iterate: function(obj) {
-
-  }
-};
