--- app/recursion.js
@@ -1,19 +1,0 @@
-exports = typeof window === 'undefined' ? global : window;
-
-exports.recursionAnswers = {
-  listFiles: function(data, dirName) {
-
-  },
-
-  permute: function(arr) {
-
-  },
-
-  fibonacci: function(n) {
-
-  },
-
-  validParentheses: function(n) {
-
-  }
-};
