--- app/regex.js
@@ -1,27 +1,0 @@
-exports = typeof window === 'undefined' ? global : window;
-
-exports.regexAnswers = {
-  containsNumber: function(str) {
-
-  },
-
-  containsRepeatingLetter: function(str) {
-
-  },
-
-  endsWithVowel: function(str) {
-
-  },
-
-  captureThreeNumbers: function(str) {
-
-  },
-
-  matchesPattern: function(str) {
-
-  },
-
-  isUSD: function(str) {
-
-  }
-};
