--- app/strings.js
@@ -1,15 +1,0 @@
-exports = typeof window === 'undefined' ? global : window;
-
-exports.stringsAnswers = {
-  reduceString: function(str, amount) {
-
-  },
-
-  wordWrap: function(str, cols) {
-
-  },
-
-  reverseString: function(str) {
-
-  }
-};
