+++ tests/runner.html
@@ -28,32 +28,14 @@
   </script>
 
   <!-- load tests and answers -->
-  <script src="/tests/app/arrays.js"></script>
-  <script src="/app/arrays.js"></script>
-  <script src="/tests/app/async.js"></script>
-  <script src="/app/async.js"></script>
-  <script src="/tests/app/bestPractices.js"></script>
-  <script src="/app/bestPractices.js"></script>
-  <script src="/tests/app/count.js"></script>
-  <script src="/app/count.js"></script>
-  <script src="/tests/app/flowControl.js"></script>
-  <script src="/app/flowControl.js"></script>
-  <script src="/tests/app/functions.js"></script>
-  <script src="/app/functions.js"></script>
-  <script src="/tests/app/logicalOperators.js"></script>
-  <script src="/app/logicalOperators.js"></script>
-  <script src="/tests/app/modules.js"></script>
-  <script src="/app/modules.js"></script>
-  <script src="/tests/app/numbers.js"></script>
-  <script src="/app/numbers.js"></script>
-  <script src="/tests/app/objects.js"></script>
-  <script src="/app/objects.js"></script>
-  <script src="/tests/app/recursion.js"></script>
-  <script src="/app/recursion.js"></script>
-  <script src="/tests/app/regex.js"></script>
-  <script src="/app/regex.js"></script>
-  <script src="/tests/app/strings.js"></script>
-  <script src="/app/strings.js"></script>
+  <script src="/tests/app/01-basics.js"></script>
+  <script src="/app/01-basics.js"></script>
+  <script src="/tests/app/02-arrays-objects.js"></script>
+  <script src="/app/02-arrays-objects.js"></script>
+  <script src="/tests/app/03-functions.js"></script>
+  <script src="/app/03-functions.js"></script>
+  <script src="/tests/app/04-concepts.js"></script>
+  <script src="/app/04-concepts.js"></script>
 
   <!-- livereload snippet -->
   <script src="http://localhost:35729/livereload.js"></script>
